Fix events 400, content save 500, and mixes auth for DJ portal
- Add ImageUrl/ImageUrls to EventVenueDto so events query can resolve venue images
- Change HeroOverlayOpacity from decimal to double (Float scalar) to fix input deserialization
- Change DJ mix mutations from RequireAdmin to RequireAuthentication so DJs can create mixes

+++ Application/DTO/SiteSettingsDTO/SiteSettingsDto.cs
@@ -16,7 +16,7 @@ namespace DJDiP.Application.DTO.SiteSettingsDTO
         public string HeroCtaLink { get; set; } = string.Empty;
         public string HeroBackgroundImageUrl { get; set; } = string.Empty;
         public string HeroBackgroundVideoUrl { get; set; } = string.Empty;
-        public decimal HeroOverlayOpacity { get; set; }
+        public double HeroOverlayOpacity { get; set; }
         public string HeroGenres { get; set; } = string.Empty;
         public string HeroLocation { get; set; } = string.Empty;
         public string HeroVibes { get; set; } = string.Empty;
